requirements.txt
fastapi==0.109.0
uvicorn[standard]==0.27.0
python-multipart==0.0.6
pillow==10.1.0
numpy==1.24.3
opencv-python==4.8.1.78
torch>=2.1.0
torchvision>=0.16.0
ultralytics==8.0.230
pydantic==2.5.0
python-dotenv==1.0.0
aiofiles==23.2.1
python-cors==4.0.0
sqlalchemy==2.0.23
pymysql==1.1.0
easyocr==1.7.0
